app/utils.py
import attackcti  # To interact with the MITRE ATT&CK framework
from pwn import asm, shellcraft, cyclic  # pwntools for payload generation


def fetch_techniques_for_actor(actor_name):
    """
    Fetch techniques associated with a given threat actor.
    Args:
        actor_name (str): Name of the threat actor.

    Returns:
        list: List of techniques associated with the threat actor.
    """
    client = attackcti.AttackClient()
    threat_actors = client.get_groups()

    # Find the specified threat actor and return their techniques
    for actor in threat_actors:
        if actor_name.lower() in actor.get('name', '').lower():
            return actor.get('techniques', [])
    return []


def validate_techniques(input_techniques, actor_techniques):
    """
    Validate and filter the techniques provided by the user.
    Args:
        input_techniques (list): Techniques provided by the user.
        actor_techniques (list): Techniques fetched from the threat actor profile.

    Returns:
        tuple: (list of valid techniques, list of invalid techniques)
    """
    valid_techniques = [tech for tech in input_techniques if tech in actor_techniques]
    invalid_techniques = [tech for tech in input_techniques if tech not in actor_techniques]
    return valid_techniques, invalid_techniques


def generate_payloads(techniques):
    """
    Generate payloads for the specified techniques.
    Args:
        techniques (list): List of valid techniques.

    Returns:
        dict: A dictionary with technique IDs as keys and payloads as values.
    """
    payloads = {}
    for technique in techniques:
        try:
            if "T1059" in technique:  # Example: Scripting
                payloads[technique] = asm(shellcraft.sh())  # Generate a shell payload
            elif "T1203" in technique:  # Example: Exploitation for Client Execution
                payloads[technique] = cyclic(100)  # Generate a cyclic pattern
            else:
                payloads[technique] = f"No payload implemented for technique {technique}."
        except Exception as e:
            payloads[technique] = f"Error generating payload: {str(e)}"
    return payloads
