requirements.txt
streamlit
attackcti
mitreattack-python
pwntools